PixelClock and Top_Module
The Pixel Clock will allows to control the refresh rate of the controller via using the clk signal to create another clock for it, as this signal can be modified, different refresh rates can be used. The Top_Module is still a WIP, and will be used to synthesize all when the moment comes, right now it was used for testing the Pixel Clock.